Add m3u specificaly for Citycable in Lausanne, FTTH
The `neptlus.m3u` file seems to differ substantially from the
actual offering in Lausanne over FTTH. This file reflects the
status of Citycable in Lausanne as of 2019-12-18.
The channel numbers are those used over DVB-T and
in the IPTV box. The channel names are exactly those
announced over DVB-T, with the exception of "TeleSwizz"
(still listed as "Be Curious TV"), "Welt" (still listed as
"Reserve 233"), "ERT World" (still "ERT 3"), and "BNT 4"
(still "BNT World"). Also, "ATV Avrupa" is misspelt as
"ATV Azrupa".
I removed from the header the xmltv file, since it only
contains a subset of the channels, and the syntax is
not understood by MythTV; this list is now
MythTV-compatible.
The list should be correct as of 2019-12-18.